modulo scopus actualizado

+++ scopus/models/DBmodel.py
@@ -1,4 +1,5 @@
 from scopus.scopusdb import Base, engine
+from sqlalchemy_utils import database_exists, create_database
 
 from scopus.models.Autores import Autores
 from scopus.models.Productos import Productos
@@ -6,7 +7,10 @@ from scopus.models.MetaDBSco import MetaDBSco
 
 
 def create_scopus_db():
-    #######################
-    Base.metadata.drop_all(engine)
-    Base.metadata.create_all(engine)
-    ###########################
+    
+    if not database_exists(engine.url):
+        create_database(engine.url)
+        Base.metadata.create_all(engine)
+    else:
+        Base.metadata.drop_all(engine)
+        Base.metadata.create_all(engine)
